requirements.txt
langchain_ollama
langchain_core
python-dotenv
streamlit
langchain_community
beautifulsoup4
bs4
pypdf
faiss-cpu
groq
langchain_groq
sentence-transformers
unstructured[pdf]